feat: adding spesific tag seeding

+++ database/migrations/2025_05_31_120554_create_mahasiswa_preferences_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('mahasiswa_preferences', function (Blueprint $table) {
             $table->string("nim");
             $table->foreign("nim")->references("nim")->on("mahasiswa");
-            $table->integer("id_tag");
+            $table->unsignedBigInteger("id_tag");
             $table->foreign("id_tag")->references("id")->on("tag");
         });
     }
